+++ src/main/java/com/lifeEgg/dto/UserAgeDTO.java
@@ -11,43 +11,19 @@ import lombok.Data;
 public class UserAgeDTO {
 	
 	@JsonIgnore
-    private int id;
+    private Long id;
 	
 	private List<Birthday> birthdays;
-
-//    public List<Birthday> getBirthdays() {
-//        return birthdays;
-//    }
-//
-//    public void setBirthdays(List<Birthday> birthdays) {
-//        this.birthdays = birthdays;
-//    }
 	
 	@Data
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Birthday {
         private Date date;
-
-//        public Date getDate() {
-//            return date;
-//        }
-//
-//        public void setDate(Date date) {
-//            this.date = date;
-//        }
     }
 
 	@Data
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class Date {
         private Integer year;
-
-//        public Integer getYear() {
-//            return year;
-//        }
-//
-//        public void setYear(Integer year) {
-//            this.year = year;
-//        }
     }
 }
